requirements.txt
# requirements.txt
requests
tqdm
python-docx
pylatexenc
python-dotenv